requirements.txt
# Core Flask & Web Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Security-Too==5.4.3
Flask-RESTX==1.3.0
Flask-Limiter==3.8.0

# Database & Caching
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
alembic==1.14.0

# Authentication & Security
authlib==1.3.2
cryptography==44.0.0
bcrypt==4.2.1
PyJWT==2.10.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1

# Environment and Configuration
python-dotenv==1.0.1

# Data Processing & Analytics
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
scipy==1.15.3
statsmodels==0.14.4
lifelines==0.30.0
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1
openpyxl==3.1.5

# Background Tasks & Monitoring
celery==5.4.0
gunicorn==23.0.0
sentry-sdk==2.19.2
prometheus-client==0.21.1

# Testing & Quality Assurance
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
selenium==4.27.1
factory-boy==3.3.1
faker==33.1.0

# Utilities & File Processing
python-dotenv==1.0.1
openpyxl==3.1.5
xlrd==2.0.1
python-magic==0.4.27
fuzzywuzzy==0.18.0
python-levenshtein==0.26.0

# Additional Dependencies for Advanced Features
click==8.1.8
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
MarkupSafe==3.0.2
WTForms==3.2.1
flask-cors==5.0.0
flask-caching==2.3.0
jsonschema==4.23.0
marshmallow==3.23.2
blinker==1.9.0
